Initial upload.
This is an example of procedurally generating vertexes and meshes, as well as procedurally generating 'constituents' which could make up the individual voxels, such as dirt, stone, etc... if I remember correctly, which could then be used to determine which texture should be wrapped around the voxel. This would allow multiple constituent parts to be present in the voxel, but interactions with those constituents would only affect the visuals if the constituent ratios changed enough to result in a different constituent being dominant. I got everything to work for display purposes, but never created actors to start pathfinding. I'm still trying to figure out the best way to path-find in a game like that instead of just going with some variation of A*.